fix(ffi): convert stream discriminants through their repr totally (#827)
The FPSS event converter widened the `Disconnected` / `Reconnecting` / `ReconnectsExhausted` reason and the `ReqResponse` result onto the C-ABI `i32` with a bare `enum as i32` cast, while the sibling reconnect-count fields already used a checked conversion. A cast on a future wider or high-bit enum repr would wrap silently to a defined-but-wrong integer handed to the C consumer.

Route both discriminants through their declared repr with the infallible `i32::from(*reason as i16)` / `i32::from(*result as u8)`. `RemoveReason` is `#[repr(i16)]` and `StreamResponseType` is `#[repr(u8)]`, so every discriminant widens losslessly and totally into the wire `i32` with no panic, no wrap, and no sentinel needed. The form is also future-proof: should a repr ever stop fitting `i32`, the conversion fails to compile rather than truncating onto the wire. A truncating `as i32` would simply wrap.

The same conversion is emitted identically across the FFI, Python, and TypeScript binding generators so the discriminant handling stays at full cross-binding parity.

+++ ffi/src/fpss_event_converter.rs
@@ -512,7 +512,7 @@ pub(crate) fn fpss_event_to_ffi(event: &thetadatadx::fpss::StreamEvent) -> FfiBu
                     event: ThetaDataDxStreamEvent {
                         kind: ThetaDataDxStreamEventKind::Disconnected,
                         disconnected: ThetaDataDxStreamDisconnected {
-                            reason: *reason as i32,
+                            reason: i32::from(*reason as i16),
                         },
                         market_value: ZERO_MARKET_VALUE,
                         ohlcvc: ZERO_OHLCVC,
@@ -800,7 +800,7 @@ pub(crate) fn fpss_event_to_ffi(event: &thetadatadx::fpss::StreamEvent) -> FfiBu
                     event: ThetaDataDxStreamEvent {
                         kind: ThetaDataDxStreamEventKind::Reconnecting,
                         reconnecting: ThetaDataDxStreamReconnecting {
-                            reason: *reason as i32,
+                            reason: i32::from(*reason as i16),
                             attempt: i32::try_from(*attempt).unwrap_or(i32::MAX),
                             delay_ms: *delay_ms,
                         },
@@ -837,7 +837,7 @@ pub(crate) fn fpss_event_to_ffi(event: &thetadatadx::fpss::StreamEvent) -> FfiBu
                     event: ThetaDataDxStreamEvent {
                         kind: ThetaDataDxStreamEventKind::ReconnectsExhausted,
                         reconnects_exhausted: ThetaDataDxStreamReconnectsExhausted {
-                            reason: *reason as i32,
+                            reason: i32::from(*reason as i16),
                             attempts: i32::try_from(*attempts).unwrap_or(i32::MAX),
                         },
                         market_value: ZERO_MARKET_VALUE,
@@ -874,7 +874,7 @@ pub(crate) fn fpss_event_to_ffi(event: &thetadatadx::fpss::StreamEvent) -> FfiBu
                         kind: ThetaDataDxStreamEventKind::ReqResponse,
                         req_response: ThetaDataDxStreamReqResponse {
                             req_id: *req_id,
-                            result: *result as i32,
+                            result: i32::from(*result as u8),
                         },
                         market_value: ZERO_MARKET_VALUE,
                         ohlcvc: ZERO_OHLCVC,
